Move all generated code in the gen sub-directory
