+++ examples/go-kit/services/session/gen/endpoints/endpoints.go
@@ -3,7 +3,7 @@ package session_endpoints
 import (
 	"fmt"
 	"github.com/go-kit/kit/endpoint"
-	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/session"
+	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/session/gen/pb"
 	context "golang.org/x/net/context"
 )
