+++ examples/go-kit/services/session/gen/pb/pb.go
