+++ examples/go-kit/services/session/gen/transports/http/http.go
@@ -8,8 +8,8 @@ import (
 
 	gokit_endpoint "github.com/go-kit/kit/endpoint"
 	httptransport "github.com/go-kit/kit/transport/http"
-	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/session"
 	endpoints "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/session/gen/endpoints"
+	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/session/gen/pb"
 )
 
 func MakeLoginHandler(ctx context.Context, svc pb.SessionServiceServer, endpoint gokit_endpoint.Endpoint) *httptransport.Server {
