+++ examples/go-kit/services/sprint/gen/endpoints/endpoints.go
@@ -3,7 +3,7 @@ package sprint_endpoints
 import (
 	"fmt"
 	"github.com/go-kit/kit/endpoint"
-	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/sprint"
+	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/sprint/gen/pb"
 	context "golang.org/x/net/context"
 )
