+++ examples/go-kit/templates/{{.File.Package}}/gen/endpoints/endpoints.go.tmpl
@@ -5,7 +5,7 @@ package {{.File.Package}}_endpoints
 import (
 	"fmt"
 	context "golang.org/x/net/context"
-	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/{{.File.Package}}"
+	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/{{.File.Package}}/gen/pb"
 	"github.com/go-kit/kit/endpoint"
 )
