+++ examples/go-kit/templates/{{.File.Package}}/gen/transports/grpc/grpc.go.tmpl
@@ -6,7 +6,7 @@ import (
         "fmt"
 
 	context "golang.org/x/net/context"
-	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/{{.File.Package}}"
+	pb "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/{{.File.Package}}/gen/pb"
 	grpctransport "github.com/go-kit/kit/transport/grpc"
 	endpoint "github.com/moul/protoc-gen-gotemplate/examples/go-kit/services/{{.File.Package}}/gen/endpoints"
 )
